refactor: one file for fetches
